requirements.txt
flask
wtforms
eventbrite
sqlalchemy
pillow
qrcode
pytz